requirements.txt
easydict==1.9
h5py==2.9.0
nltk==3.4.1
numpy==1.16.3
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
python-dateutil==2.8.0
PyYAML==5.1
six==1.12.0
tensorboardX==1.6
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
comet-ml
easydict
numpy
python-dotenv
pytorch-lightning>=0.7.4
PyYAML
